requirements.txt
# Core Application Dependencies
gradio==4.8.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Machine Learning
tensorflow==2.15.0
numpy==1.24.3
scikit-learn==1.3.2

# Image Processing
opencv-python-headless==4.8.1.78
Pillow==10.1.0

# Data Handling
pandas==2.1.4

# Visualization
matplotlib==3.8.2
seaborn==0.13.0

# API & Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0
PyYAML==6.0.1

# Kaggle Dataset
kagglehub==0.3.8
